requirements.txt
backports.functools-lru-cache==1.5
certifi==2018.4.16
chardet==3.0.4
cycler==0.10.0
idna==2.6
itchat==1.3.10
jieba==0.39
kiwisolver==1.0.1
matplotlib==2.2.2
numpy==1.14.3
Pillow==5.1.0
pyparsing==2.2.0
pypng==0.0.18
PyQRCode==1.2.1
python-dateutil==2.7.3
pytz==2018.4
requests==2.18.4
six==1.11.0
snownlp==0.12.3
urllib3==1.22
wordcloud==1.4.1
